correction v1.8

+++ src/backend/Approx/Butterworth.py
@@ -58,8 +58,8 @@ class Butterworth(object):
             else:
                 self.order = order
         elif self.type == "Band Pass":
-            order, wo = signal.buttord([float(2 * np.pi * fpMin), float(2 * np.pi * fpMax)],
-                                            [float(2 * np.pi * faMin),float(2 * np.pi * faMax)],
+            order, wo = signal.buttord([2 * np.pi * fpMin, 2 * np.pi * fpMax],
+                                            [2 * np.pi * faMin,2 * np.pi * faMax],
                                             Ap, Aa, analog=True)
             if Nmin is not None and Nmin > order:
                 self.order = Nmin
